trying to use saner types
	* x11_windowsys.h, x11_windows.c, x11_main.c, x11_keys.c,
	windows.c, views.c, undo.c, server-dl.c, redisplay.c,
	mac_windows.m, mac_runloop.m, mac_main.m, mac_keys.m,
	mac_internal.h, keys.c, jadeclient.c, jade_subrs.h, jade.h,
	gtk_select.c, gtk_main.c, gtk_keys.c, gtk_jade.c, gtk_defs.h,
	glyphs.c, files.c, faces.c, extent.c, editcommands.c, edit.h,
	buffers.c: trying to use saner types -- no u_{char,int,long}
	and no shorts. Use uint8_t and uint32_t where it makes sense.

+++ src/gtk_jade.c
@@ -74,7 +74,7 @@ static bool new_window_no_frame = FALSE;
 static bool new_window_no_show = FALSE;
 
 /* Timestamp from a recent event */
-u_long gtk_jade_last_event_time;
+unsigned long gtk_jade_last_event_time;
 
 DEFSYM (gtk_jade_new_hook, "gtk-jade-new-hook"); /*
 ::doc:gtk-jade-new-hook::
@@ -443,7 +443,7 @@ static gint
 gtk_jade_input_event (GtkWidget *widget, GdkEvent *event)
 {
     GtkJade *jade;
-    u_long code = 0, mods = 0;
+    unsigned long code = 0, mods = 0;
     bool redisplay = FALSE;
     g_return_val_if_fail (widget != NULL, FALSE);
     g_return_val_if_fail (GTK_IS_JADE (widget), FALSE);
@@ -757,7 +757,7 @@ focus_out_callback (GtkWidget *widget, GdkEvent *ev, gpointer data)
 
 /* The only thing necessary in W is the font stuff (I think) */
 GtkJade *
-sys_new_window(Lisp_Window *oldW, Lisp_Window *w, short *dims)
+sys_new_window(Lisp_Window *oldW, Lisp_Window *w, int *dims)
 {
     unsigned int x = -1, y = -1, width = 80, height = 24;
     GtkWidget *frame = 0;
